+++ eventmesh-connector-rocketmq/src/main/java/org/apache/eventmesh/connector/rocketmq/common/Constants.java
@@ -1,0 +1,27 @@
+package org.apache.eventmesh.connector.rocketmq.common;
+
+public class Constants {
+    public static final String BROADCAST_PREFIX = "broadcast-";
+
+    public final static String CONSUMER_GROUP_NAME_PREFIX = "ConsumerGroup-";
+
+    public final static String PRODUCER_GROUP_NAME_PREFIX = "ProducerGroup-";
+
+    public static final String PROPERTY_MESSAGE_TIMEOUT = "TIMEOUT";
+
+    public static final String PROPERTY_MESSAGE_KEYS = "KEYS";
+
+    public static final String PROPERTY_MESSAGE_QUEUE_ID = "QUEUE_ID";
+
+    public static final String PROPERTY_MESSAGE_QUEUE_OFFSET = "QUEUE_OFFSET";
+
+    public static final String PROPERTY_MESSAGE_DESTINATION = "DESTINATION";
+
+    public static final String PROPERTY_MESSAGE_MESSAGE_ID = "MESSAGE_ID";
+
+    public static final String PROPERTY_MESSAGE_BORN_HOST = "BORN_HOST";
+
+    public static final String PROPERTY_MESSAGE_BORN_TIMESTAMP = "BORN_TIMESTAMP";
+
+    public static final String PROPERTY_MESSAGE_STORE_HOST = "STORE_HOST";
+}
