--- eventmesh-connector-rocketmq/src/main/resources/META-INF/services/com.webank.eventmesh.api.consumer.MeshMQPushConsumer
@@ -1,1 +1,0 @@
-com.webank.eventmesh.connector.rocketmq.consumer.RocketMQConsumerImpl
