+++ eventmesh-connector-rocketmq/src/main/resources/META-INF/services/org.apache.eventmesh.api.consumer.MeshMQPushConsumer
@@ -1,0 +1,1 @@
+org.apache.eventmesh.connector.rocketmq.consumer.RocketMQConsumerImpl
