--- eventmesh-connector-rocketmq/src/test/resources/META-INF/services/io.openmessaging.MessagingAccessPoint
@@ -1,1 +1,0 @@
-com.webank.eventmesh.connector.rocketmq.MessagingAccessPointImpl
