+++ eventmesh-connector-rocketmq/src/test/resources/META-INF/services/org.apache.io.openmessaging.MessagingAccessPoint
@@ -1,0 +1,1 @@
+org.apache.eventmesh.connector.rocketmq.MessagingAccessPointImpl
