+++ eventmesh-connector-rocketmq/src/test/resources/META-INF/services/org.apache.io.openmessaging.producer.Producer
@@ -1,0 +1,1 @@
+org.apache.eventmesh.connector.rocketmq.producer.ProducerImpl
